sound: Refactor editor classes.
