--- VisualPinball.Engine/VPT/ISoundEmitter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7129feb5a774383458adb84d3b08fabe
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
