--- VisualPinball.Unity/VisualPinball.Unity.Editor/Inspectors/MechSoundDrawer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 884fb5b527309ef489e8b27aa9e4809d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
