--- VisualPinball.Unity/VisualPinball.Unity.Editor/Inspectors/MechSoundsInspector.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8dca340821f92c74e8cf97cd3fff29df
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
