--- VisualPinball.Unity/VisualPinball.Unity.Editor/VPT/Sounds.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8dc2c04b7ef683d4a8076a16ebddc298
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
