--- VisualPinball.Unity/VisualPinball.Unity.Editor/VPT/Sounds/SoundInspector.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bace99bbc8f020f49b66c0ce06780514
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
