+++ VisualPinball.Unity/VisualPinball.Unity/Sound/SoundAsset.cs.meta
