+++ VisualPinball.Unity/VisualPinball.Unity/Sound/SoundTrigger.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 82f302a40f4f455d819e782d1f1a6127
+timeCreated: 1677678720
