--- VisualPinball.Unity/VisualPinball.Unity/VPT/Sounds.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f028dd712c2180947bba5a587be6fca6
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
