--- VisualPinball.Unity/VisualPinball.Unity/VPT/Sounds/GameSounds.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 941dfc7e839dd5f4194835c1c47fb2c4
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
