--- VisualPinball.Unity/VisualPinball.Unity/VPT/Sounds/MechSounds.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4372796933d0ef74fa75d5603c72aab8
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
